requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.22.0.dev0=pypi_0
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.0.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
authlib=1.2.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
bitsandbytes=0.41.1=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.2.0=pypi_0
cryptography=41.0.3=pypi_0
dataclasses-json=0.5.14=pypi_0
decorator=5.1.1=pypi_0
einops=0.6.1=pypi_0
filelock=3.12.2=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.32=pypi_0
google-api-core=2.11.1=pypi_0
google-api-python-client=2.96.0=pypi_0
google-auth=2.22.0=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
googleapis-common-protos=1.60.0=pypi_0
greenlet=2.0.2=pypi_0
httplib2=0.22.0=pypi_0
huggingface-hub=0.16.4=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
jsonschema=4.19.0=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
langchain=0.0.260=pypi_0
langsmith=0.0.20=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lxml=4.9.3=pypi_0
markdown-it-py=3.0.0=pypi_0
marshmallow=3.20.1=pypi_0
mdurl=0.1.2=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
numexpr=2.8.5=pypi_0
numpy=1.25.2=pypi_0
oauthlib=3.2.2=pypi_0
openai=0.27.8=pypi_0
openapi-schema-pydantic=1.2.4=pypi_0
openssl=1.1.1v=h7f8727e_0
pandas=2.0.3=pypi_0
peft=0.5.0.dev0=pypi_0
pillow=9.5.0=pypi_0
pip=23.2.1=py310h06a4308_0
protobuf=4.24.0=pypi_0
psutil=5.9.5=pypi_0
pyarrow=12.0.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydantic=1.10.12=pypi_0
pydeck=0.8.0=pypi_0
pygments=2.16.1=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.1.1=pypi_0
pyre-extensions=0.0.29=pypi_0
python=3.10.0=h12debd9_5
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
pytz=2023.3=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
regex=2023.8.8=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.5.2=pypi_0
rpds-py=0.9.2=pypi_0
rsa=4.9=pypi_0
safetensors=0.3.2=pypi_0
setuptools=68.0.0=py310h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlalchemy=2.0.19=pypi_0
sqlite=3.41.2=h5eee18b_0
streamlit=1.25.0=pypi_0
tenacity=8.2.2=pypi_0
tiktoken=0.4.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.3.2=pypi_0
tqdm=4.66.0=pypi_0
transformers=4.32.0.dev0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=4.3.1=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.16=pypi_0
validators=0.20.0=pypi_0
watchdog=3.0.0=pypi_0
weaviate-client=3.22.1=pypi_0
wheel=0.38.4=py310h06a4308_0
xformers=0.0.20=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
